requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.4
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
